style(#24): rename AppRoutes
