+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import './App.css';
-import AppRoutes from "./AppRoutes";
+import AppRoutes from "./routes/AppRoutes";
 
 const App: React.FC = () => {
   const [load, setLoad] = useState(true);
