+++ src/routes/AppRoutes.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes, Navigate } from "react-router-dom";
 
-import About from "./page/about";
-import Contact from "./components/Contact/Contact";
-import Resume from "./page/resume";
-import Preloader from "../src/components/Pre";
-import Blog from "./page/blog";
-import Portfolio from "./page/portfolio";
-import ScrollToTop from "./components/ScrollToTop";
-import Docs from "./page/docs";
+import About from "../page/about";
+import Contact from "../components/Contact/Contact";
+import Resume from "../page/resume";
+import Preloader from "../../src/components/Pre";
+import Blog from "../page/blog";
+import Portfolio from "../page/portfolio";
+import ScrollToTop from "../components/ScrollToTop";
+import Docs from "../page/docs";
 
 const routes = [
   { path: "/", component: About },
